add parse hz code file code

+++ Haze/Haze/Memory/Header/MemoryPage.h
@@ -9,8 +9,8 @@ struct MemoryPageInfo
 {
 	std::unique_ptr<MemoryBlock> HeadBlock;
 
-	uint32 PageByteSize;
-	uint32 UnitSize;
+	uint64 PageByteSize;
+	uint64 UnitSize;
 	
 	std::unique_ptr<MemoryPage> NextPage;
 };
@@ -20,7 +20,7 @@ class MemoryPage
 public:
 	friend class GarbageCollection;
 
-	MemoryPage(uint32 PageByteSize, uint32 BlockUnitSize);
+	MemoryPage(uint64 PageByteSize, uint64 BlockUnitSize);
 	~MemoryPage();
 
 	void* TryAlloca(uint64 Size);
